stdshaders: Fix double -> float downcast warnings

+++ materialsystem/stdshaders/spritecard.cpp
@@ -416,8 +416,8 @@ SHADER_DRAW
 
 		if ( bZoomSeq2 )
 		{
-			float flZScale=1.0/(params[ZOOMANIMATESEQ2]->GetFloatValue());
-			float C0[4]={ (float)(0.5*(1.0+flZScale)), flZScale, 0, 0 };
+			float flZScale=1.0f/(params[ZOOMANIMATESEQ2]->GetFloatValue());
+			float C0[4]={ 0.5f*(1.0f+flZScale), flZScale, 0, 0 };
 			pShaderAPI->SetVertexShaderConstant( VERTEX_SHADER_SHADER_SPECIFIC_CONST_7, C0,
 				ssize(C0)/4 );
 		}
